feat: add select-account functionality

+++ services/oauth-service/src/handler/logout/handle-refresh-session-logout.ts
@@ -1,31 +1,35 @@
+import { Client, LogoutSession } from "../../entity";
+import { ServerError } from "@lindorm-io/errors";
 import { ServerKoaContext } from "../../types";
-import { LogoutSession } from "../../entity";
 import { createLogoutToken } from "../token";
-import { handleConsentSessionOnLogout } from "./handle-consent-session-on-logout";
 
 export const handleRefreshSessionLogout = async (
   ctx: ServerKoaContext,
   logoutSession: LogoutSession,
+  client: Client,
 ): Promise<void> => {
   const {
     axios: { axiosClient },
-    cache: { clientCache },
     repository: { refreshSessionRepository },
   } = ctx;
 
+  if (!logoutSession.confirmedLogout.refreshSessionId) {
+    throw new ServerError("Invalid session state", {
+      debug: { confirmedLogout: logoutSession.confirmedLogout },
+    });
+  }
+
   const refreshSession = await refreshSessionRepository.find({
-    id: logoutSession.sessionId,
+    id: logoutSession.confirmedLogout.refreshSessionId,
   });
 
-  const client = await clientCache.find({ id: refreshSession.clientId });
+  const { token } = createLogoutToken(ctx, client, refreshSession);
 
-  const { token: logoutToken } = await createLogoutToken(ctx, client, refreshSession);
-
-  await axiosClient.post(client.logoutUri, {
-    body: { logoutToken },
-  });
-
-  await handleConsentSessionOnLogout(ctx, client, refreshSession);
+  if (client.backChannelLogoutUri) {
+    await axiosClient.post(client.backChannelLogoutUri, {
+      body: { logoutToken: token },
+    });
+  }
 
   await refreshSessionRepository.destroy(refreshSession);
 };
